Rename client.js and delete server.js
*Server.js is no longer needed. Now the bot is
using AJAX instead sockets

*Adjust the paths of the import in index.html
