+++ public/index.html
@@ -5,8 +5,7 @@
 	<title>Bodo IT Help Desk</title>
 	<link rel="stylesheet" type="text/css" href="index.css">
 	<script src="https://code.jquery.com/jquery-3.1.1.min.js"></script>
-    <script src="/socket.io/socket.io.js"></script>
-    <script src="client.js"></script> 
+    <script src="./client.js"></script> 
 </head>
 <body>
 	<div class="chatbox">
@@ -14,7 +13,7 @@
 		</div>
 		<div class="chat-form">
 			<textarea rows="5" id="message" ></textarea>
-			<button id="send" onclick="sendToServer()" >Send</button>
+			<button id="send" onclick="sendToServer()">Send</button>
 		</div>
 	</div>
 </body>
